+++ CHANGELOG.rst
@@ -32,9 +32,13 @@ Each individual change should have a link to the pull request after the descript
 
 Changed
 ^^^^^^^
-
 - turned on DOC lints for remaining classes `#547 <https://github.com/azukds/tubular/issues/547>_`
 - turned on D lints for remaining classes `#115 <https://github.com/azukds/tubular/issues/115>_`
+- copied optimisations/improvements from #484 into MedianImputer
+- feat: added `to_json` method for MeanResponseTransformer `#549 <https://github.com/azukds/tubular/issues/549>_`
+- feat: added `to_json` method for DatetimeInfoExtractor `#555 <https://github.com/azukds/tubular/issues/555>_`
+- feat: added `to_json` method for `DatetimeSinusoidCalculator` `#556 <https://github.com/azukds/tubular/issues/556>`_
+- feat: added a new transformer called DatetimeComponentExtractor
 
 
 2.3.0 (18/11/2025)
@@ -61,6 +65,7 @@ Changed
 - feat: added `to_json` method for `SetValueTransformer` `#542 <https://github.com/azukds/tubular/issues/542>`_
 - feat: added 'to_json' method for GroupRareLevelsTransformer '#548 <https://github.com/azukds/tubular/issues/548>'
 - removed SeparatorColumnMixin `#562 <https://github.com/azukds/tubular/issues/562>_`
+- added a new transformer called DatetimeComponentExtractor
 
 2.1.0 (30/10/25)
 ------------------